fix post scope in post-panel component
