+++ src/app/sidebar/feed/feed.component.ts
@@ -47,16 +47,6 @@ export class FeedComponent implements AfterViewInit {
     this.updatePosts();
   }
 
-  openDeleteDialog(): void {
-    const dialogRef = this.dialog.open(ApprovedDialogComponent, {
-      width: '250px',
-    });
-
-    dialogRef.afterClosed().subscribe(result => {
-      console.log('The dialog was closed');
-    });
-  }
-
   ngAfterViewInit() {
     this.markers = [this.augustMarker, this.septemberMarker, this.octoberMarker, this.novemberMarker,
                     this.decemberMarker, this.januaryMarker, this.februaryMarker, this.marchMarker,
