+++ src/app/sidebar/feed/post-panel/post-panel.component.html
@@ -5,7 +5,7 @@
         <div class="inner-container">
           <mat-panel-title>{{post.title}}</mat-panel-title>
           <mat-panel-description>Currently I am {{panelOpenState ? 'open' : 'closed'}}</mat-panel-description>
-          <mat-icon matTooltip="Sent to: All high schoolers">people_outline</mat-icon>
+          <mat-icon matTooltip="All high schoolers">people_outline</mat-icon>
         </div>
         <div class="teaser-text" *ngIf=!panelOpenState>
           SENIORS: JOSTEN'S WILL RETURN this Friday, October 19 in the main gym lobby during lunch to collect Graduation item orders.
