+++ src/app/sidebar/feed/post-panel/post-panel.component.ts
@@ -1,5 +1,7 @@
 import {Component, Input, OnInit} from '@angular/core';
 import {Post} from '../../../post';
+import {ApprovedDialogComponent} from '../approved-dialog/approved-dialog.component';
+import {MatDialog} from '@angular/material';
 
 @Component({
   selector: 'app-post-panel',
@@ -11,7 +13,19 @@ export class PostPanelComponent implements OnInit {
   panelOpenState = false;
   @Input() post: Post;
 
-  ngOnInit() {
+  constructor(public dialog: MatDialog) {
+  }
+
+  ngOnInit() { }
+
+  openDeleteDialog(): void {
+    const dialogRef = this.dialog.open(ApprovedDialogComponent, {
+      width: '250px',
+    });
+
+    dialogRef.afterClosed().subscribe(result => {
+      console.log('The dialog was closed');
+    });
   }
 
 }
